BookShop/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
blessings==1.7
bpython==0.21
certifi==2021.5.30
charset-normalizer==2.0.6
curtsies==0.3.5
cwcwidth==0.1.4
Django==3.2.7
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
psycopg2-binary==2.9.1
Pygments==2.10.0
python-decouple==3.5
python-slugify==5.0.2
pytz==2021.1
pyxdg==0.27
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
